# schemas/openapi/repository-api.yaml
openapi: 3.1.0
info:
  title: Civics Research Repository API
  version: 0.1.0
  description: Typed backend API contract for public research discovery, datasets, map overlays, and accessibility evidence.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: https://api.civics-research-repository.local/api
    description: Local development API host alias
security: []
...
paths:
...
  /admin/corpus/scale:
    post:
      tags: [Admin]
      operationId: startCorpusProfileScale
      summary: Start guarded asynchronous growth and activation of a supported corpus profile.
      description: >-
        Starts the durable harvest/snapshot/projection/evidence workflow. FEDERATED_100K is the
        currently supported growth target; retained publisher metadata is reused when already present.
      parameters:
        - name: profile
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/CorpusProfile'
      responses:
        '202':
          description: Scale operation accepted; poll the reindex progress endpoint for completion.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CorpusProfileActivationProgress'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /admin/corpus/scale/evidence:
    get:
      tags: [Admin]
      operationId: getCorpusScaleEvidence
      summary: Verify the current evidence chain for a named corpus profile without mutating state.
      parameters:
        - name: profile
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/CorpusProfile'
      responses:
        '200':
          description: Read-only corpus evidence report and any detected violations.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CorpusScaleEvidenceReport'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
...
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Conflict:
      description: The requested operator action conflicts with current runtime state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    CorpusProfile:
      type: string
      description: >-
        Stable local corpus profile. FEDERATED_10K/100K/1M describe target federated metadata counts
        in addition to the curated repository records; FULL has no fixed count.
      enum: [CURATED_DEMO, FEDERATED_10K, FEDERATED_100K, FEDERATED_1M, FULL]
...
    CorpusProfileActivationPhase:
      type: string
      enum:
        - IDLE
        - PREPARING
        - HARVESTING
        - SNAPSHOTTING
        - PROJECTING
        - VERIFYING
        - CAPTURING_EVIDENCE
        - COMPLETED
        - FAILED
    CorpusProfileActivationProgress:
      type: object
      required:
        - phase
        - processedDocuments
        - percentComplete
        - updatedAt
        - elapsedMs
        - message
      properties:
        operationId:
          type: [string, 'null']
        profile:
          oneOf:
            - $ref: '#/components/schemas/CorpusProfile'
            - type: 'null'
        phase:
          $ref: '#/components/schemas/CorpusProfileActivationPhase'
        processedDocuments:
          type: integer
          format: int64
          minimum: 0
        totalDocuments:
          type: [integer, 'null']
          format: int64
          minimum: 0
        percentComplete:
          type: integer
          minimum: 0
          maximum: 100
        startedAt:
          type: [string, 'null']
          format: date-time
        updatedAt:
          type: string
          format: date-time
        completedAt:
          type: [string, 'null']
          format: date-time
        elapsedMs:
          type: integer
          format: int64
          minimum: 0
        documentsPerSecond:
          type: [number, 'null']
          format: double
          minimum: 0
        message:
          type: string
    CorpusScaleEvidenceReport:
      type: object
      required:
        - profile
        - valid
        - retainedFederatedRecordCount
        - currentProjectionObjectCount
        - targetParity
        - storageEvidencePresent
        - violations
      properties:
        profile:
          $ref: '#/components/schemas/CorpusProfile'
        valid:
          type: boolean
        targetFederatedRecordCount:
          type: [integer, 'null']
          format: int64
          minimum: 0
        retainedFederatedRecordCount:
          type: integer
          format: int64
          minimum: 0
        activeProfile:
          oneOf:
            - $ref: '#/components/schemas/CorpusProfile'
            - type: 'null'
        activationProjectionObjectCount:
          type: [integer, 'null']
          format: int64
          minimum: 0
        activationProjectionId:
          type: [string, 'null']
          pattern: '^[0-9a-f]{64}$'
        currentProjectionObjectCount:
          type: integer
          minimum: 0
        currentProjectionId:
          type: [string, 'null']
          pattern: '^[0-9a-f]{64}$'
        targetParity:
          type: boolean
        storageEvidencePresent:
          type: boolean
        storageProjectionObjectCount:
          type: [integer, 'null']
          format: int64
          minimum: 0
        storageRetainedFederatedCount:
          type: [integer, 'null']
          format: int64
          minimum: 0
        storageProjectionId:
          type: [string, 'null']
          pattern: '^[0-9a-f]{64}$'
        storageCapturedAt:
          type: [string, 'null']
          format: date-time
        violations:
          type: array
          items:
            type: string
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable error code.
        message:
          type: string
          description: Human-readable summary safe to show in UI.
        details:
          type: array
          items:
            type: string
          description: Optional validation or diagnostic detail lines.
        traceId:
          type: string
          description: Correlation id for logs and support, echoed from X-Trace-Id when supplied.
...